% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{Cluster}
\alias{Cluster}
\title{Cluster}
\usage{
Cluster(nij = 4, age = c(2, 1.4, 1, 0.7, 0.5), delta = 2,
  colo = c("blue", "lightblue", "darkgreen", "green"), icon = 8 * cod,
  seed = 12345, ...)
}
\arguments{
\item{nij}{symbols per size}

\item{age}{vector of sizes, default c(2,1.4,1,0.7,0.5)}

\item{delta}{standardized mean difference}

\item{colo}{vector of extreme colors in both clusters, default c("blue","lightblue","darkgreen","green")}

\item{icon}{symbol as a matrix}

\item{seed}{for random number generator}

\item{...}{input to function lines}
}
\description{
Cluster
}
\examples{
\dontrun{
Cluster(delta=1.5)
Cluster(7,c(1.41,1.3,1.21,1.13,1.06,1,0.95,0.91),1)
Cluster(icon=5*fir)
Cluster(icon=3*worm,colo=c("lightgrey","white","lightgray","cyan"))
Cluster(icon=5*flower,colo=c("blue","lightblue","red","pink"))
Cluster(7,c(1.41,1.3,1.21,1.13,1.06,1,0.95,0.91),2, icon=5*man,colo=c("blue","lightblue","red","white"))
Cluster(icon=5*bike,colo="no",col="blue",lwd=3)
}
}
