% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{Forest}
\alias{Forest}
\title{Title}
\usage{
Forest(ni = 30, path = 5, nrow = 0, near = 3, icon = fir,
  sizeratio = 1.4, colri = c("green", "lightgreen"),
  colle = c("darkgreen", "green"), ...)
}
\arguments{
\item{ni}{trees per stand}

\item{path}{gap between stands}

\item{nrow}{number of rows of trees, if 0, uniformly random}

\item{near}{factor to the distance between symbols}

\item{icon}{symbol as a matrix}

\item{sizeratio}{factor by which symbols in right stand are larger}

\item{colri}{extreme colors of right stand}

\item{colle}{extreme colors of left stand}

\item{...}{input to function lines()}
}
\value{
plot of two groups of symbols in rows and columns or uniformly distributed
}
\description{
Title
}
\examples{
\dontrun{
Forest()
Forest(ni=28,nrow=5,near=5)
Forest(path=2,ni=50)
Forest(ni=28,nrow=5,near=5,icon=bush)
Forest(ni=15,nrow=5,near=-2,icon=wheat-3,path=11,
       colle=c("yellow","gold"),colri=c("lightgreen","yellow"),col="brown")
Forest(ni=12,nrow=4,near=-2,icon=0.8*bike,path=11,
       colle=c("white","white"),colri=c("white","white"),col="blue",lwd=5)
Forest(ni=15,nrow=5,near=-2,icon=arabidopsis/2.2,path=11,
       colle=c("lightgreen","green"),colri=c("green","darkgreen"),col="darkgreen")
}
}
