% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{Gradient}
\alias{Gradient}
\title{Gradient}
\usage{
Gradient(age = c(2, 1.4, 1, 0.7, 0.5), ni = 9, near = 1, a = 0, b = 1,
  icon = fir, colo = c("darkgreen", "green"), seed = 12345, ...)
}
\arguments{
\item{age}{vector of sizes}

\item{ni}{number of symbols per size}

\item{near}{factor applied to distance between symbols}

\item{a}{intercept}

\item{b}{slope}

\item{icon}{symbol as a matrix}

\item{colo}{vector of colors at the end of the color gradient spanning length(age) colors, default is c("darkgreen","green")}

\item{seed}{for random number generator}

\item{...}{input to function lines()}
}
\value{
plot of a gradient
}
\description{
Gradient
}
\examples{
\dontrun{
Gradient()
Gradient(ni=19,near=0.5)
Gradient(c(1.41,1.3,1.21,1.13,1.06,1),29,icon=rbind(bush,fir-2))
Gradient(a=100,b=-0.6,icon=cod)
Gradient(exp((10:1)/5-4),near=0.5,icon=8*cod,colo = c("lightblue","cyan"))
Gradient(c(1.41,1.3,1.21,1.13,1.06,1),near=0.5,icon=wheat,colo = c("lightgreen","gold"),col="brown")
Gradient(icon=worm,colo = c("lightgray","cyan"),col="grey")
Gradient(icon=bike,colo="no",col="blue",lwd=3)
}
}
