% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan}
\alias{plan}
\title{Title}
\usage{
plan(nij = 5, r = 2, c = 2, re = 1, ce = 1, ie = 1, icon = rat,
  colo = c("darkgrey", "white"), seed = 1, ...)
}
\arguments{
\item{nij}{group size}

\item{r}{rows}

\item{c}{columns}

\item{re}{row effect}

\item{ce}{column effect}

\item{ie}{interaction effect}

\item{icon}{symbol as a matrix}

\item{colo}{extremes of colors}

\item{seed}{for random number generator}

\item{...}{input to function lines()}
}
\value{
plot of (r x c)-plan
}
\description{
Title
}
\examples{
\dontrun{
plan()
plan(4,3,2,0.5,2,3)
plan(6,icon=cod*2.5)
plan(6,,,icon=4*fir,colo="darkgreen")
plan(6,,,icon=5*man,colo="darkgreen")
plan(9,,,icon=2*arabidopsis,colo="green")
plan(icon=4*zebrafish,colo="n",col="darkgrey")
plan(9,r=2,c=3,re=0.1,ce=0.2,ie=0,icon=3*arabidopsis,colo=c("darkgreen","yellow"),col="green")
}
}
