% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmplot.R
\name{plotglm}
\alias{plotglm}
\title{Title}
\usage{
plotglm(icon = fir, iex = NULL, colo = c("darkgreen", "green"), rot = 0,
  str = 1, data, ...)
}
\arguments{
\item{icon}{symbol as a matrix}

\item{iex}{iconsize scaling factor}

\item{colo}{extreme colors}

\item{rot}{rotation in degrees for every symbol}

\item{str}{horizontal stretch factor applied to every symbol}

\item{data}{dataframe of simulated data, e.g. by function rglm}

\item{...}{input to function lines()}
}
\value{
plots scatterplot
}
\description{
Title
}
\examples{
\dontrun{
plotglm(data=rglm(s=3),str=0.5)
plotglm(icon=E.coli,colo=c("gray","yellow"),col="darkgrey",
   data = rglm(dist="poisson",xes=list(x1=c(10,170),x2=3,x3=9,x4=2)))
# Regressoren bedeuten: x-Achse, Farbe, Größe, Drehung
daten <- rglm(dist="poisson",
               xes=list(x1=c(10,20),x2=2,x3=c(0.7,2.5),x4=c(-2,6)),
              beta=c(1,0.5,2,.03,1,0.7))
plotglm(icon=E.coli,colo=c("gray","green"),col="darkgrey",rot=10, iex=.2, str=0.3, data = daten)
daten <- rglm(n=12,dist="binomial", xes=list(x1=c(7,14),x2=1,x3=1,x4=1), beta=c(0,0.032,0,0,0,0))
daten[,5] <- daten[,1]
daten[,4] <- daten[,2]/10
plotglm(icon=zebrafish,colo=c("n"),col="black",rot=pi, iex=.04, str = 1.3, data = daten)
est <- glm(y~x.x1,family='binomial',data=daten)
new <- data.frame(x.x1=(1:200)/10)
odd <- exp(predict(est,newdata=new))
P <- odd/(1+odd)
lines(new$x.x1,P)
abline(v=-est$coefficients[1]/est$coefficients[2],h=c(0,0.5,1),lty=2)
}
}
