% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tplot.R
\name{Tplot}
\alias{Tplot}
\title{Tplot}
\usage{
Tplot(x, N = sum(x[, 1]), nc = max(apply(as.matrix(row.names(x)), 1,
  stringr::str_length)), ns = 1, verbose = F, icon = "bar", space = 1,
  icol = "blue", col = "lightblue", cut = FALSE,
  xlab = "Häufigkeit (\%)", bcol = "white", border = par("fg"),
  s.x = 1, s.y = 1, seed = 12345, ...)
}
\arguments{
\item{x}{matrix with labels in rownames and absolute frequencies in x[,1]}

\item{N}{total number (denominator), if not sum(x[,1])}

\item{nc}{number of characters of labels}

\item{ns}{scal of labels}

\item{verbose}{if TRUE, frequencies appear in labels}

\item{icon}{two-column matrix of coordinates or "bar" for barplot}

\item{space}{factor, usually 0.5 < 1, applied to x-dimension of icons for spacing}

\item{icol}{icon outline color}

\item{col}{icon fill color}

\item{cut}{shall partial icons be displayed, if representing more than one observation?}

\item{xlab}{xlab}

\item{bcol}{bar color}

\item{border}{bar and axis line color}

\item{s.x}{SD to jitter icons horizontally in max(relative frequency)/50}

\item{s.y}{SD to jitter icons vertically in dim(x)[1]/200}

\item{seed}{seed for jittering}

\item{...}{graphics parameters handed to call of barplot}

\item{y}{vector containing dependent variable}
}
\value{
Plots the specified barplot using jittered icons.
}
\description{
Tplot
}
\examples{
 x <- matrix(1:3,3,1)
rownames(x) <-c("eins","zwei","dreivierfünfsechssieben\\n achtneunzehnelfzwölf")
Tplot(x,45,nc=50,ns=2)
Tplot(x,nc=30,ns=2,icon=fir,icol="darkgreen",col="green",bcol="#FAFAFA", border="#EEEEEE",cut=TRUE,cex.axis=1.7,cex.lab=1.7)
Tplot(x,nc=30,ns=2,icon=cod[,c(2,1)],icol="blue",col="lightblue",border="white")
Tplot(x,nc=30,ns=1.5,icon=human,icol="gray20",col="mistyrose",border="lavender")
x <- matrix(c(3,4,9,17,57),ncol=1)
rownames(x)<- c("Ziege","Schaf","Rind","Schwein","Huhn")
Tplot(x,nc=5,ns=1,icon=bone[,c(2,1)],space=0.7,icol="grey50",bcol="grey90",border="grey80")
}
