% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battleship.R
\name{battleship}
\alias{battleship}
\title{Battleship plot}
\usage{
battleship(y, cens = 0, x = 1, dis = "normal", xprop = T, space = 1,
  his = NA, binz = 0, err = F, loc = "none", prec = "2sd",
  connect = F, box = F, vio = F, dot = NA, icon = "loop",
  fill = FALSE, ship = T, plank = 2, dtr = F,
  fillcol = shape::shadepalette(length(unique(x)), "lightblue", IMBSgreen.),
  elicol = shape::shadepalette(length(unique(x)), "blue", "darkgreen"),
  errcol = shape::shadepalette(length(unique(x)), IMBSorange, "brown"),
  concol = shape::shadepalette(length(as.matrix(connect)[1, ]), IMBSorange,
  "brown"), hiscol = shape::shadepalette(length(unique(x)), "darkgreen",
  "darkblue"), boxcol = shape::shadepalette(length(unique(x)), "brown",
  IMBSorange), viocol = shape::shadepalette(length(unique(x)), IMBSorange,
  "brown"), shipcol = shape::shadepalette(length(unique(x)), IMBSorange,
  "brown"), quant = c(0.025, 0.25, 0.5, 0.75, 0.975), cline = FALSE,
  clc = "white", whisker = 0, confluence = "Tufte", trials = 0,
  start = list(m), mix = NA, genmix = F, df = 4, kern = "gaussian",
  sdsk = sqrt(var(y))/log(length(y)), ...)
}
\arguments{
\item{y}{vector containing dependent variable}

\item{cens}{vector of length(y), is > 0 where y is right-censored, is < 0 where left-censored, 0 elsewhere implemented for dist "lognormal", "normal", "gamma" or "weibull"}

\item{x}{factor, same length as y}

\item{dis}{distribution, "lognormal", "normal", "2normal", "t", "beta","gamma", "weibull","mixnorm", "poisson", "binomial", "betabinomial"}

\item{xprop}{x-axis proportional to group sizes (joint distribution)? recommended, if dot=TRUE}

\item{space}{multiplicative spacing in x-direction}

\item{his}{plot histogram or probability function?}

\item{binz}{numbers of bins for dotplots and histograms}

\item{err}{draw error bar? Not drawn when data are censored}

\item{loc}{location of error bar or error bars, if vector: "mean", "geomean"}

\item{prec}{whiskers of error bar, "se", "sd", "2se", "2sd", applied to logarithms, if loc="geomean".}

\item{connect}{whether arithmetic or geometric means according to loc are connected, as scalar logical value or matrix of integers <= number of x levels with one column per polygon}

\item{box}{draw boxplots?}

\item{vio}{draw something like a violin plot?}

\item{dot}{plot dotplot? logical vector of length 1 or number of groups}

\item{icon}{either a character name of predefined icons or a matrix with coordinates, when character vector or list of matrices of length = levels of x, then icon by x, when list of length = levels of x of character vectors of number of ys per level, or list of lists of matrices then single icons sorted first by x second by y}

\item{fill}{fill icons? If sum(abs(cens)) > 0, then censored values are empty, others filled}

\item{ship}{plot battleship? logical vector of length 1 or number of x levels}

\item{plank}{line width for plotting ships and quantiles, if vector then for each ship}

\item{dtr}{density trace residual?}

\item{fillcol}{color the icons are filled with, when scalar then all the same, when vector of length = levels of x, then color by x, when list of length = levels of x of vectors of number of icons per level, color of single icons sorted first by x second by y}

\item{elicol}{color the icons are bordered with, similar to fillcol}

\item{errcol}{color the error bars are drawn in, similar to boxcol}

\item{concol}{color the connecting lines are drawn in, similar to boxcol}

\item{hiscol}{histogram color, if vector then for each histogram}

\item{boxcol}{boxplot color, if vector then for each boxplot}

\item{viocol}{violin color, if vector then for each violin}

\item{shipcol}{ship color, if vector then for each ship}

\item{quant}{vector of quantile for which bars are drawn across}

\item{cline}{plot line(s) through the middle at x label?}

\item{clc}{color(s) of line(s) through the middle}

\item{whisker}{battleships are drawn to the whisker quantile, default = 0.5 / n_i}

\item{confluence}{"Tukey" outlined bars, "Tufte" thin bars, "Duerr" circles}

\item{trials}{size parameter of binomial distribution(s)}

\item{start}{list of starting values for mixture parameters}

\item{mix}{number of normals in a mixture}

\item{genmix}{whether mixtures differ in proportions only}

\item{df}{degrees of freedom of t-distribution(s)}

\item{kern}{kernel of density trace for dtr or violin}

\item{sdsk}{smoothness parameter of density trace for dtr or violin}

\item{...}{}
}
\value{
Prints tuning parameters binz ans space and returns histograms of y for values of x. Plots the specified plot.
}
\description{
Battleship plot
}
\examples{
x <- rep(c(1:4),c(100,100,100,100)*.1); y <- x+rnorm(40,x/2)
 battleship(y=y,x=x,dtr=FALSE,fill=TRUE,xprop=FALSE, ship=TRUE, dis="normal", plank=1, cline=TRUE, space=1.7, dot=TRUE, binz=13, icon=rabbit, fillcol=c("brown","gold","lightgray","white"))
}
